se crea el crud completo de productos
se crea el crud completo de productos sin validación

+++ backend/src/main/java/com/cafe/repositoy/ProductoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cafe.entity.ProductoEntity;
 
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer> {
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Long> {
 
 }
